Добавил genre mpa

+++ src/main/java/ru/yandex/practicum/filmorate/dao/dto/FilmDto.java
@@ -14,7 +14,7 @@ public class FilmDto {
     private String name;
     private String description;
     private LocalDate releaseDate;
-    private int duration;
+    private Long duration;
     private Set<Integer> genreIds;
     private Integer mpa;
 }
